requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
Faker==26.0.0
pytest==8.2.0
pytest-cov==5.0.0
requests==2.32.3
selenium==4.21.0